requirements.txt
click==6.6
Flask==0.11.1
flask-mongoengine==0.8
Flask-Script==2.0.5
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mongoengine==0.10.6
pymongo==3.3.1
python-dateutil==2.6.0
requests==2.12.1
six==1.10.0
Werkzeug==0.11.11
wheel==0.24.0
WTForms==2.1
